Move original files back to lib/
