+++ lib/cbor.js
