+++ lib/commented.js
