+++ lib/decoder.js
