+++ lib/diagnose.js
