+++ lib/encoder.js
