+++ lib/map.js
