+++ test/commented.ava.js
@@ -1,9 +1,9 @@
 'use strict'
 
-const cbor = require('../src/cbor')
+const cbor = require('../')
 const test = require('ava')
 const cases = require('./cases')
-const utils = require('../src/utils')
+const utils = require('../lib/utils')
 const NoFilter = require('nofilter')
 
 function testAll(t, list) {
