+++ test/diagnose.ava.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const NoFilter = require('nofilter')
-const cbor = require('../src/cbor')
+const cbor = require('../')
 const test = require('ava')
 const cases = require('./cases')
